Cleanup ntdll
