+++ MonkeyWorks.csproj
@@ -5,7 +5,7 @@
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
     <PackageReadmeFile>README.md</PackageReadmeFile>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Version>2.0.5</Version>
+    <Version>2.0.6</Version>
     <Authors>0xBadJuju</Authors>
     <Company>NetSPI</Company>
     <Copyright>BSD-3-Clause</Copyright>
@@ -32,6 +32,10 @@
 
   <ItemGroup>
     <Compile Remove="Misc.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Remove="Unmanaged\Headers\Ntpsapi.cs~RFc34d20a.TMP" />
   </ItemGroup>
 
 </Project>
