+++ Unmanaged/Headers/Ntpsapi.cs
@@ -1,41 +1,75 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 
-using USHORT = System.UInt16;
-using WORD = System.UInt16;
-
-using DWORD = System.UInt32;
-using ULONG = System.UInt32;
-
-using QWORD = System.UInt64;
-using ULONGLONG = System.UInt64;
-using LARGE_INTEGER = System.UInt64;
-
-using PVOID = System.IntPtr;
-using LPVOID = System.IntPtr;
-using DWORD_PTR = System.IntPtr;
-using SIZE_T = System.IntPtr;
-using PWSTR = System.IntPtr;
-
 namespace MonkeyWorks.Unmanaged.Headers
 {
-    public sealed class ntpsapi
+    // So this is a non-public file that we guess these items came from, so yeah, sorry
+    // https://processhacker.sourceforge.io/doc/ntpsapi_8h_source.html
+    // https://www.geoffchappell.com/studies/windows/km/ntoskrnl/source/inc/ntpsapi.htm
+    // If I can't find another place for structs used by ntdll, they are going here
+
+    public sealed class Ntpsapi
     {
-        //Process Security and Access Rights
-        //https://msdn.microsoft.com/en-us/library/windows/desktop/ms684880%28v=vs.85%29.aspx?f=255&MSPPError=-2147217396
-        internal const DWORD PROCESS_ALL_ACCESS = 0;
-        internal const DWORD PROCESS_CREATE_PROCESS = 0x0080;
-        internal const DWORD PROCESS_CREATE_THREAD = 0x0002;
-        internal const DWORD PROCESS_DUP_HANDLE = 0x0040;
-        internal const DWORD PROCESS_QUERY_INFORMATION = 0x0400;
-        internal const DWORD PROCESS_QUERY_LIMITED_INFORMATION = 0x1000;
-        internal const DWORD PROCESS_SET_INFORMATION = 0x0200;
-        internal const DWORD PROCESS_SET_QUOTA = 0x0100;
-        internal const DWORD PROCESS_SUSPEND_RESUME = 0x0800;
-        internal const DWORD PROCESS_TERMINATE = 0x0001;
-        internal const DWORD PROCESS_VM_OPERATION = 0x0008;
-        internal const DWORD PROCESS_VM_READ = 0x0010;
-        internal const DWORD PROCESS_VM_WRITE = 0x0020;
-        internal const DWORD SYNCHRONIZE = 0x00100000;
+        [StructLayout(LayoutKind.Sequential, Pack = 1)]
+        public struct _INITIAL_TEB
+        {
+            public IntPtr StackBase;
+            public IntPtr StackLimit;
+            public IntPtr StackCommit;
+            public IntPtr StackCommitMax;
+            public IntPtr StackReserved;
+        }
+
+        [StructLayout(LayoutKind.Sequential, Pack = 1)]
+        public struct _PROCESS_ACCESS_TOKEN
+        {
+            public IntPtr hToken;
+            public IntPtr hThread;
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct SYSTEM_HANDLE_INFORMATION
+        {
+            [MarshalAs(UnmanagedType.U8)]
+            public ulong NumberOfHandles;
+            public SYSTEM_HANDLE_TABLE_ENTRY_INFO[] Handles;
+        }
+
+        [StructLayout(LayoutKind.Sequential, Pack = 1)]
+        public struct SYSTEM_HANDLE_TABLE_ENTRY_INFO
+        { // Information Class 16
+            public int ProcessID;
+            public byte ObjectTypeNumber;
+            public byte Flags; // 0x01 = PROTECT_FROM_CLOSE, 0x02 = INHERIT
+            public ushort Handle;
+            public int Object_Pointer;
+            public uint GrantedAccess;
+        }
+
+        [StructLayout(LayoutKind.Sequential, Pack = 1)]
+        public struct _SYSTEM_HANDLE_TABLE_ENTRY_INFO
+        {
+            public ushort UniqueProcessId;
+            public ushort CreatorBackTraceIndex;
+            public byte ObjectTypeIndex;
+            public byte HandleAttributes;
+            public ushort HandleValue;
+            public IntPtr Object;
+            public ulong GrantedAccess;
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct _THREAD_BASIC_INFORMATION
+        {
+            [MarshalAs(UnmanagedType.U4)]
+            public uint ExitStatus;
+            public IntPtr TebBaseAddress;
+            public Winnt._LIST_ENTRY ClientId;
+            public IntPtr AffinityMask;
+            [MarshalAs(UnmanagedType.I4)]
+            public int Priority;
+            [MarshalAs(UnmanagedType.I4)]
+            public int BasePriority;
+        }
     }
 }
