+++ Unmanaged/Libraries/PInvoke/ntdll.cs
@@ -151,7 +151,7 @@ namespace MonkeyWorks.Unmanaged.Libraries
         [DllImport("ntdll.dll", SetLastError = true)]
         public static extern uint NtQueryInformationProcess(
             IntPtr ProcessHandle,
-            _PROCESS_INFORMATION_CLASS ProcessInformationClass,
+            ProcessThreadsApi._PROCESS_INFORMATION_CLASS ProcessInformationClass,
             IntPtr ProcessInformation,
             uint ProcessInformationLength,
             ref uint ReturnLength
@@ -168,7 +168,7 @@ namespace MonkeyWorks.Unmanaged.Libraries
         [DllImport("ntdll.dll", SetLastError = true)]
         public static extern uint NtSetInformationProcess(
             IntPtr ProcessHandle,
-            _PROCESS_INFORMATION_CLASS ProcessInformationClass,
+            ProcessThreadsApi._PROCESS_INFORMATION_CLASS ProcessInformationClass,
             ref uint ProcessInformation,
             uint ProcessInformationLength
         );
@@ -176,8 +176,8 @@ namespace MonkeyWorks.Unmanaged.Libraries
         [DllImport("ntdll.dll", SetLastError = true)]
         public static extern uint NtSetInformationProcess(
             IntPtr ProcessHandle,
-            _PROCESS_INFORMATION_CLASS ProcessInformationClass,
-            ref _PROCESS_ACCESS_TOKEN ProcessInformation,
+            ProcessThreadsApi._PROCESS_INFORMATION_CLASS ProcessInformationClass,
+            ref Ntpsapi._PROCESS_ACCESS_TOKEN ProcessInformation,
             uint ProcessInformationLength
         );
 
@@ -216,81 +216,5 @@ namespace MonkeyWorks.Unmanaged.Libraries
         public static extern uint RtlNtStatusToDosError(
             uint Status
         );
-
-        [StructLayout(LayoutKind.Sequential, Pack = 1)]
-        public struct _PROCESS_ACCESS_TOKEN
-        {
-            public IntPtr hToken;
-            public IntPtr hThread;
-        }
-
-        [Flags]
-        public enum _PROCESS_INFORMATION_CLASS
-        {
-            ProcessBasicInformation = 0,
-            ProcessQuotaLimits,
-            ProcessIoCounters,
-            ProcessVmCounters,
-            ProcessTimes,
-            ProcessBasePriority,
-            ProcessRaisePriority,
-            ProcessDebugPort,
-            ProcessExceptionPort,
-            ProcessAccessToken,
-            ProcessLdtInformation,
-            ProcessLdtSize,
-            ProcessDefaultHardErrorMode,
-            ProcessIoPortHandlers,
-            ProcessPooledUsageAndLimits,
-            ProcessWorkingSetWatch,
-            ProcessUserModeIOPL,
-            ProcessEnableAlignmentFaultFixup,
-            ProcessPriorityClass,
-            ProcessWx86Information,
-            ProcessHandleCount,
-            ProcessAffinityMask,
-            ProcessPriorityBoost,
-            ProcessDeviceMap,
-            ProcessSessionInformation,
-            ProcessForegroundInformation,
-            ProcessWow64Information,
-            ProcessImageFileName,
-            ProcessLUIDDeviceMapsEnabled,
-            ProcessBreakOnTermination,
-            ProcessDebugObjectHandle,
-            ProcessDebugFlags,
-            ProcessHandleTracing,
-            ProcessIoPriority,
-            ProcessExecuteFlags,
-            ProcessTlsInformation,
-            ProcessCookie,
-            ProcessImageInformation,
-            ProcessCycleTime,
-            ProcessPagePriority,
-            ProcessInstrumentationCallback,
-            ProcessThreadStackAllocation,
-            ProcessWorkingSetWatchEx,
-            ProcessImageFileNameWin32,
-            ProcessImageFileMapping,
-            ProcessAffinityUpdateMode,
-            ProcessMemoryAllocationMode,
-            ProcessGroupInformation,
-            ProcessTokenVirtualizationEnabled,
-            ProcessConsoleHostProcess,
-            ProcessWindowInformation,
-            MaxProcessInfoClass,
-            ProcessProtectionInformation = 61
-        }
-
-        [StructLayout(LayoutKind.Sequential, Pack = 1)]
-        public struct _PROCESS_BASIC_INFORMATION
-        {
-            public IntPtr Reserved1;
-            public IntPtr PebBaseAddress;
-            public IntPtr AffinityMask;
-            public IntPtr BasePriority;
-            public UIntPtr UniqueProcessId;
-            public IntPtr Reserved3;
-        }
     }
 }
